Fix REDUCTION and ALLOCATE clause.
In the lexer, the predefined parameter, such as "inscan" and "min", should be differed from similar expression like "inscantest" or "mintest".
This issue is solved by limiting the postfix of predefined parameter.
Specifically, "inscan" goes to "inscan/{blank}*,". It means word "inscan" that followed any number of space and a comma. This pattern differs "inscan" from the expression "inscantest".
Here, "/" means the postfix is concerned, but it stays in the unprocessed text.