For adding clause object to directive IR, we now
use addOpenMPClause as the single entry from the grammar to add clause. In the IR implementation, there is a static function for adding a clause to a directive object, for each clause class that extends the base OpenMPClause class. So when adding a new clause and if we need to create a new subclass of OpenMPClause, we will need to add that static function in the class for trying adding the parsed clause to the directive. The function only takes clause parameters (not the clause object itself) since the function needs to check whether it needs to create a clause object or return an existing clause that has the same parameter.