Implement initial version of test driver.
If no argument is given, it will test "tests/parallel.test", which is a sample testing file.
Additional testing file could be created based on this file as template.

If any case failed, the details including original input, wrong output and correct output will be printed.
At last, a summary will be printed.