Fix the incorrect default ALLOCATOR in ALLOCATE clause.
If the enum that indicates there's no allocator is not specified, the allocator enum value will be set to 0 by default, which is actuall OMPC_ALLOCATE_ALLOCATOR_default.
Therefore, we have to specify it to OMPC_ALLOCATE_ALLOCATOR_unknown.

This bug probably happens to ALLOCATOR and USESALLOCATOR clauses as well. We need to check and fix them.