Implement normalization and fix a few minor bugs.
Since the size of expression list is supposed to be small, brute force searching is used for now.

In the switch-case statement, a default case should be added to handle unexpected cases. After each case, the code should be always contained with a pair of curly bracket to avoid errors after future changes.