Implement error handler and fix bugs in the lexer.
If the parser encounter an error, it will not terminate immediately. Instead, it will output the error information and return the parsing result at that moment.

When an error happened, the buffer and states in the lexer have to be
reset manually. Otherwise, the outdated data will be used by the next round of parsing.

In the lexer, there was a bug that when some error appears during scanning the lexer used to either try to access an empty buffer leading to segment fault, or not send the last expression string.
Therefore, while hitting the EOF rule, the lexer checks whether there's a cached string waiting to be sent. If so, return the string as token.