Add a vector to record the original order of clauses
A vector of clause pointers is added to the OpenMPDirective class.
A int index is added to the OpenMPClause class. Its initial value is -1. If the new clause could be normalized, an updated existing clause will be returned and its index must not be -1.
In the other hand, the index of new created clause is -1. We set the index to the size of current size of ordered clause vector.

For the clauses with post-normalization, only the latest clause could be removed from the clause map. Therefore, we can just pop out the last clause in the vector when this kind of normalization happens.